+++ color_picker/color_picker_loop.py
@@ -3,204 +3,38 @@
 from pathlib import Path
 from argparse import ArgumentParser
 from typing import List, Dict, Any, Tuple
-from itertools import product
 from typing import Optional
-from threading import Thread
 import json
-import copy
-import yaml
 import numpy as np
 import os, shutil
 import matplotlib.pyplot as plt
 
-#wei
-from rpl_wei import WEI
 
 #solvers utils
+#For extracting colors from each plate
 from tools.plate_color_analysis import get_colors_from_file
+#Different possible solvers for the color_picker problem
 from solvers.bayes_solver import BayesColorSolver
 from solvers.evolutionary_solver import EvolutionaryColorSolver
 from solvers.aggressive_genetic_solver import AggroColorSolver
 from funcx import FuncXExecutor
 from datetime import datetime
-import cv2
+#For publishing to RPL Portal
 from tools.publish import publish_iter
-from tools.log_info import get_log_info
-from skopt import Optimizer
+#For creating a payload that the OT2 will accept from the solver output
+from tools.volumes_to_payload import convert_volumes_to_payload
+#For running WEI flows
+from tools.run_flow import run_flow
+#for measuring the three mixed colors for calibration and for 
+# ensuring the target color is in the right color space
+from tools.calibrate import calibrate
+#For constructing the plots for each run
+from tools.create_visuals import create_visuals, create_target_plate
 from datetime import datetime
 from gladier import GladierBaseClient, generate_flow_definition, GladierBaseTool
 
 #DEF TODO:what does it mean?
 MAX_PLATE_SIZE = 96
-
-def new_plate():
-    well_rows = ["A", "B", "C", "D", "E", "F", "G", "H"]
-    well_cols = [str(elem) for elem in range(1, 13)]
-    well_names = ["".join(elem) for elem in product(well_rows, well_cols)]
-    return well_names
-
-def convert_volumes_to_payload(volumes: List[List[float]], curr_wells_used: List[Any]) -> Tuple[Dict[str, Any], List[Any]]:
- 
-    well_names = new_plate()
-    curr_wells_available = copy.copy(well_names)
-    if len(volumes) <= len(well_names) and len(volumes) <= MAX_PLATE_SIZE - len(curr_wells_used):
-        for i in curr_wells_used:
-            curr_wells_available.remove(i)
-        
-    r_vol, g_vol, b_vol = [], [], []
-    dest_wells = []
-    for color, well in zip(volumes, curr_wells_available):
-    # for color, well in zip(volumes, well_names):
-        r, g, b = color
-        r_vol.append(r)
-        g_vol.append(g)
-        b_vol.append(b)
-        dest_wells.append(well)
-        curr_wells_used.append(well)
-    return {
-        "red_volumes": r_vol,
-        "green_volumes": g_vol,
-        "blue_volumes": b_vol,
-        "destination_wells": dest_wells,
-    }, curr_wells_used
-
-class ThreadWithReturnValue(Thread):
-    
-    def __init__(self, group=None, target=None, name=None,
-                 args=(), kwargs={}, Verbose=None):
-        Thread.__init__(self, group, target, name, args, kwargs)
-        self._return = None
-
-    def run(self):
-        if self._target is not None:
-            self._return = self._target(*self._args,
-                                                **self._kwargs)
-        
-    def join(self, *args):
-        Thread.join(self, *args)
-        return self._return
-
-def wei_run_flow(wf_file_path, payload):
-    wei_client = WEI(wf_file_path)
-    run_info = wei_client.run_workflow(payload=payload)
-    #print(run_info)
-    return run_info
-def run_flow(protocol, payload, steps_run):
-                iter_thread=ThreadWithReturnValue(target=wei_run_flow,kwargs={'wf_file_path':protocol,'payload':payload})
-                iter_thread.run()
-                run_info = iter_thread._return
-                run_dir = run_info["run_dir"]
-                t_steps_run = get_log_info(run_dir, protocol)
-                steps_run.append(t_steps_run)
-                return steps_run, run_info
-
-def create_target_plate(plate_volumes: List[List[float]],
-                        colors: List[List[int]]):
-        norms = []
-        for i in plate_volumes:
-            norms.append(i/sum(i))
-        target_plate = np.array(norms)@np.array(colors)
-        target_plate = target_plate.tolist()
-        return target_plate
-
-    
-def calibrate(target_color: List[int], 
-              curr_wells_used: List[str], 
-              loop_protocol: str, 
-              exp_folder: Path, 
-              plate_max_volume: float,
-              steps_run: List[Dict[str, Any]],
-              pop_size: int
-              )-> (List[List[int]], List[int], List[str], List[Dict[str, Any]]):
-    """Performs a calibration run of the color picker system
-@Inputs: 
-    target_color: RGB Color the solver is attempting to match
-    curr_wells_used: The wells that have so far been filled with liquid by the
-    loop_protocol: Workcell protocol used to mix colors
-    exp_folder: Folder where the experiment data is saved
-    steps_run: The steps run in the workflow so far
-@Outputs:
-    colors: colors that are combined linearly to visualize experimental colors
-    target_color: RGB Color captured from camera image of actual plate based on the ratio calculated from the input RGB values
-    steps_run: """
-    plate_volumes = np.array([[0.98, 0.01, 0.01], [0.01, 0.98, 0.01], [0.01, 0.01, 0.98], np.array(target_color)/sum(target_color)])*plate_max_volume
-    payload, curr_wells_used = convert_volumes_to_payload(plate_volumes, curr_wells_used)
-    payload['use_existing_resources'] = False 
-    steps_run, run_info = run_flow(loop_protocol, payload, steps_run) 
-    fname = "final_image.jpg" #image"+str(ot2_iter) +".jpg"
-    img_path = run_info["run_dir"]/ "results" / fname
-    plate_colors_ratios = get_colors_from_file(img_path)[1]
-    plate_colors_ratios = {a:b[::-1] for a,b in plate_colors_ratios.items()}  
-    current_plate = []
-    wells_used = []
-    for well in payload["destination_wells"]:
-        color = plate_colors_ratios[well]
-        wells_used.append(well)
-        current_plate.append(color)
-    target_color = current_plate[3]
-    target_color = target_color.tolist()
-    colors = current_plate[0:3]
-    t = np.asarray(colors)
-    colors = t.tolist()
-    color_image = cv2.resize( np.asarray([colors]).astype(np.uint8), [pop_size*50, 50], interpolation = cv2.INTER_NEAREST)
-    plt.imsave(exp_folder/"results"/"mixed_colors.png", color_image/255)
-    return colors, target_color, curr_wells_used, steps_run
-#Creates the plots for the color picker system
-#@Inputs:
-    # target_plate: Expected colors output by solver
-    # current_plate: Colors measured from camera
-    # exp_folder: Folder where the experiment data is saved
-    # current_iter: Current iteration of the experiment
-    # target_color: Color the solver is attempting to match
-    # cur_best_color: The color mixed so far that best matches the target color
-    # pop_size: The number of wells used per experiment run 
-#@Outputs:
-    #None
-def create_visuals(target_plate: List[List[int]],
-                   current_plate: List[List[int]],
-                   exp_folder: Path,
-                   current_iter: int,
-                   target_color: List[int],
-                   cur_best_color: List[int],
-                   pop_size: int,
-                   diffs: List[List[float]], 
-                   solver_out_dim: int
-    ) -> None:
- 
-    solver.plot_diffs(diffs, exp_folder)
-    plt.ion()
-    f, axarr = plt.subplots(2, 2)
-    # set figure size to 10x10
-    f.set_figheight(10)
-    f.set_figwidth(10)
-    graph_vis = np.asarray(target_plate)
-    graph_vis = graph_vis.reshape(*solver_out_dim)
-    plate_vis = np.asarray(current_plate)
-    plate_vis = plate_vis.reshape(*solver_out_dim)
-    exp_p = np.asarray([graph_vis]).astype(np.uint8)
-    exp_p = cv2.resize( np.asarray([graph_vis]).astype(np.uint8), [pop_size*50, 50], interpolation = cv2.INTER_NEAREST)
-    axarr[0][0].imshow(exp_p)
-    axarr[0][0].set_title("Experiment plate")
-    axarr[1][0].imshow([plate_vis])
-    axarr[1][0].set_title("Real plate")
-    axarr[0][1].imshow([[target_color]])
-    axarr[0][1].set_title("Target Color")
-    axarr[1][1].imshow([[cur_best_color]])
-    axarr[1][1].set_title("Experiment best color")
-    f.suptitle("PAUSING HERE TO MOVE THE PLATE")
-    f.canvas.draw()
-    f.canvas.flush_events()
-    plt.pause(0.001)
-    plt.savefig(exp_folder/"results"/"run_summary.png", dpi=300)
-    exp_p = np.asarray([graph_vis]).astype(np.uint8)
-    exp_p = cv2.resize( np.asarray([graph_vis]).astype(np.uint8), [pop_size*50, 50], interpolation = cv2.INTER_NEAREST)
-    plt.imsave(exp_folder/"results"/str("run_" + str(current_iter) + "_expected.png"), exp_p)
-    np.asarray([plate_vis]).astype(np.uint8)
-    real_p = cv2.resize( np.asarray([plate_vis]).astype(np.uint8), [pop_size*50, 50], interpolation = cv2.INTER_NEAREST)
-    plt.imsave(exp_folder/"results"/("run_" + str(current_iter) + "_measured.png"), real_p)
-    plt.imsave(exp_folder/"results"/"target_color.png", np.asarray([[target_color]])/255),
-    plt.imsave(exp_folder/"results"/"best_color.png", np.asarray([[cur_best_color ]])/255)
-
     
 def run(
     exp_type: str,
@@ -267,6 +101,7 @@ def run(
     time_to_best = str(start - start)
     
     while num_exps + pop_size <= exp_budget:
+        new_run = {}
         steps_run = []
         log_line = 0
         run_dir = ""
@@ -373,7 +208,7 @@ def run(
         current_iter += 1
         num_exps += pop_size
         ##Plot review
-        create_visuals(target_plate, current_plate, exp_folder, current_iter, target_color, cur_best_color, pop_size, diffs, solver_out_dim)
+        create_visuals(target_plate, current_plate, exp_folder, current_iter, target_color, cur_best_color, pop_size, diffs, solver_out_dim, solver)
         runs = []
         report = {}
         if (exp_folder/"results"/"exp_data.txt").is_file():
